+++ test/IntegrationTests/PluginsTests.cs
@@ -14,6 +14,7 @@
 // limitations under the License.
 // </copyright>
 
+using System.Linq;
 using System.Threading.Tasks;
 using IntegrationTests.Helpers;
 using Xunit;
@@ -35,6 +36,11 @@ public class PluginsTests : TestHelper
         using var collector = await MockSpansCollector.Start(Output);
         SetExporter(collector);
         collector.Expect("MyCompany.MyProduct.MyLibrary");
+#if NETFRAMEWORK
+        collector.Expect("OpenTelemetry.HttpWebRequest", span => span.Attributes.Any(att => att.Key == "example.plugin"));
+#else
+        collector.Expect("OpenTelemetry.Instrumentation.Http", span => span.Attributes.Any(att => att.Key == "example.plugin"));
+#endif
 
         SetEnvironmentVariable("OTEL_DOTNET_AUTO_PLUGINS", "TestApplication.Plugins.Plugin, TestApplication.Plugins, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null");
         RunTestApplication();